add increase allowance button, separate files (#480)
* add increase allowance button, separate files

* rename file
